# Move proof targets to the heap