I actually meant to inline the producing functions initially, but this seems to result in better performance:
```
Before:
chia/proofs/single/1x   time:   [710.02 ms 713.94 ms 718.19 ms]
                        thrpt:  [1.3924  elem/s 1.4007  elem/s 1.4084  elem/s]
chia/proofs/parallel/8x time:   [567.42 ms 574.66 ms 581.51 ms]
                        thrpt:  [13.757  elem/s 13.921  elem/s 14.099  elem/s]
After:
chia/proofs/single/1x   time:   [681.83 ms 683.49 ms 685.20 ms]
                        thrpt:  [1.4594  elem/s 1.4631  elem/s 1.4666  elem/s]
chia/proofs/parallel/8x time:   [554.84 ms 558.87 ms 562.89 ms]
                        thrpt:  [14.212  elem/s 14.315  elem/s 14.419  elem/s]
```